refactor: Move indicator classes to separate files and create 'indicators' as a submodule
